requirements.txt
fake-bpy-module-latest==20230308
